DOM functions moved to DOMupdate file
